requirements.txt

# Core Libraries
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0

# Machine Learning & Embeddings
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.26.0

# System Utilities
psutil==5.9.5
requests==2.31.0

# Optional: For Documentation
python-multipart==0.0.9
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
mypy==1.8.0
black==24.1.1
